Windows: change specs.Process.Args to a single string
This is necessary because Windows natively only supports an argument
string, not an argument array. There are some useful Windows command
lines that are not possible to express using just an array of arguments
that are combined using syscall.EscapeArg. It seems better to express this
clearly in the spec rather than adding a bool for marking that arguments
do not need escaping.

Signed-off-by: John Starks <jostarks@microsoft.com>