Windows: Clean up argument escaping and revert behavior
This cleans up the argument escaping to only appear in Windows-specific
code, and it reverts the behavior of the entrypoint to never be escaped.
Post-1.11, we will again escape the entrypoint (to match Linux behavior),
but this will require several test updates and may break existing
Dockerfiles or other scripts.

Signed-off-by: John Starks <jostarks@microsoft.com>